[SERVICE-571] Rename StoredSetting
