+++ src/provider/model/StoredSetting.ts
